+++ sentry-apache-http-client-5/src/main/java/io/sentry/transport/apache/ApacheHttpClientTransport.java
@@ -14,7 +14,6 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.zip.GZIPOutputStream;
 import org.apache.hc.client5.http.async.methods.SimpleHttpRequest;
 import org.apache.hc.client5.http.async.methods.SimpleHttpRequests;
@@ -36,15 +35,14 @@ public final class ApacheHttpClientTransport implements ITransport {
   private final @NotNull RequestDetails requestDetails;
   private final @NotNull CloseableHttpAsyncClient httpclient;
   private final @NotNull RateLimiter rateLimiter;
-  private final @NotNull AtomicInteger currentlyRunning;
-  final ReusableCountLatch latch = new ReusableCountLatch();
+  private final @NotNull ReusableCountLatch currentlyRunning;
 
   public ApacheHttpClientTransport(
       final @NotNull SentryOptions options,
       final @NotNull RequestDetails requestDetails,
       final @NotNull CloseableHttpAsyncClient httpclient,
       final @NotNull RateLimiter rateLimiter) {
-    this(options, requestDetails, httpclient, rateLimiter, new AtomicInteger());
+    this(options, requestDetails, httpclient, rateLimiter, new ReusableCountLatch());
   }
 
   ApacheHttpClientTransport(
@@ -52,7 +50,7 @@ public final class ApacheHttpClientTransport implements ITransport {
       final @NotNull RequestDetails requestDetails,
       final @NotNull CloseableHttpAsyncClient httpclient,
       final @NotNull RateLimiter rateLimiter,
-      final @NotNull AtomicInteger currentlyRunning) {
+      final @NotNull ReusableCountLatch currentlyRunning) {
     this.options = Objects.requireNonNull(options, "options is required");
     this.requestDetails = Objects.requireNonNull(requestDetails, "requestDetails is required");
     this.httpclient = Objects.requireNonNull(httpclient, "httpclient is required");
@@ -69,9 +67,7 @@ public final class ApacheHttpClientTransport implements ITransport {
       final SentryEnvelope filteredEnvelope = rateLimiter.filter(envelope, hint);
 
       if (filteredEnvelope != null) {
-        currentlyRunning.incrementAndGet();
-        latch.increment();
-        System.out.println("After increment " + latch.getCount());
+        currentlyRunning.increment();
 
         try (final ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
             final GZIPOutputStream gzip = new GZIPOutputStream(outputStream)) {
@@ -89,7 +85,9 @@ public final class ApacheHttpClientTransport implements ITransport {
           }
 
           if (options.getLogger().isEnabled(DEBUG)) {
-            options.getLogger().log(DEBUG, "Currently running %d requests", currentlyRunning.get());
+            options
+                .getLogger()
+                .log(DEBUG, "Currently running %d requests", currentlyRunning.getCount());
           }
 
           httpclient.execute(
@@ -97,7 +95,6 @@ public final class ApacheHttpClientTransport implements ITransport {
               new FutureCallback<SimpleHttpResponse>() {
                 @Override
                 public void completed(SimpleHttpResponse response) {
-                  currentlyRunning.decrementAndGet();
                   if (response.getCode() != 200) {
                     options
                         .getLogger()
@@ -111,24 +108,19 @@ public final class ApacheHttpClientTransport implements ITransport {
                       rateLimits != null ? rateLimits.getValue() : null,
                       retryAfter != null ? retryAfter.getValue() : null,
                       response.getCode());
-                  latch.decrement();
-                  System.out.println("After decreemnt " + latch.getCount());
+                  currentlyRunning.decrement();
                 }
 
                 @Override
                 public void failed(Exception ex) {
-                  currentlyRunning.decrementAndGet();
                   options.getLogger().log(ERROR, "Error while sending an envelope", ex);
-                  latch.decrement();
-                  System.out.println("After decreemnt " + latch.getCount());
+                  currentlyRunning.decrement();
                 }
 
                 @Override
                 public void cancelled() {
-                  currentlyRunning.decrementAndGet();
                   options.getLogger().log(WARNING, "Request cancelled");
-                  latch.decrement();
-                  System.out.println("After decreemnt " + latch.getCount());
+                  currentlyRunning.decrement();
                 }
               });
         } catch (Exception e) {
@@ -143,10 +135,9 @@ public final class ApacheHttpClientTransport implements ITransport {
   @Override
   public void flush(long timeoutMillis) {
     try {
-      System.out.println("Flush " + latch.getCount());
-      latch.waitTillZero(timeoutMillis, TimeUnit.MILLISECONDS);
+      currentlyRunning.waitTillZero(timeoutMillis, TimeUnit.MILLISECONDS);
     } catch (InterruptedException e) {
-      options.getLogger().log(SentryLevel.ERROR, "Failed to flush events");
+      options.getLogger().log(SentryLevel.ERROR, "Failed to flush events", e);
     }
   }
 
@@ -163,6 +154,6 @@ public final class ApacheHttpClientTransport implements ITransport {
   }
 
   private boolean isSchedulingAllowed() {
-    return currentlyRunning.get() < options.getMaxQueueSize();
+    return currentlyRunning.getCount() < options.getMaxQueueSize();
   }
 }